refactor: split preprocess module and test module into several modules
